[#210] Move tests to integrationtests
