+++ iceoryx2-ffi/ffi/Cargo.toml
@@ -30,3 +30,4 @@ iceoryx2-ffi-macros = { workspace = true }
 
 [dev-dependencies]
 iceoryx2-bb-testing = { workspace = true }
+generic-tests = { workspace = true }
